+++ Changelog.md
@@ -16,6 +16,7 @@ This document tracks all notables changes to the Level Up app.
 ### Fixed
 
 - Make the app work in landscape mode
+- Change input field cursor color to match the input text so it is easier to see
 
 ---